requirements.txt
alembic==0.9.8
blinker==1.4
click==6.7
config==0.3.9
configparser==3.5.0
dominate==2.3.1
Flask==0.12.2
Flask-Admin
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gunicorn
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
psycopg2==2.7.4
PyJWT==1.6.0
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2013.7
requests==2.5.1
sendgrid
six==1.11.0
SQLAlchemy==1.2.4
urllib3==1.22
virtualenv==15.1.0
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.1
WTForms-Components==0.10.3
